pyproject.toml
[project]
name = "pyspark-pipeline-composer"
version = "0.1.0"
description = "A minimal framework for composing PySpark structured streaming pipelines."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "pyspark>=4.1.1",
    "delta-spark>=4.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "databricks"
url = "https://pypi-proxy.dev.databricks.com/simple"
explicit = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
